refactor: move vm hooks into view model folder
