+++ packages/shared-components/src/room-list/RoomListHeaderView/RoomListHeaderView.tsx
@@ -9,8 +9,7 @@ import React, { type JSX } from "react";
 import { IconButton, H1 } from "@vector-im/compound-web";
 import ComposeIcon from "@vector-im/compound-design-tokens/assets/web/icons/compose";
 
-import { type ViewModel } from "../../viewmodel/ViewModel";
-import { useViewModel } from "../../useViewModel";
+import { type ViewModel, useViewModel } from "../../viewmodel";
 import { Flex } from "../../utils/Flex";
 import { useI18n } from "../../utils/i18nContext";
 import { ComposeMenuView, OptionMenuView, SpaceMenuView } from "./menu";
