+++ packages/shared-components/src/room-list/RoomListSearchView/RoomListSearchView.tsx
@@ -12,8 +12,7 @@ import SearchIcon from "@vector-im/compound-design-tokens/assets/web/icons/searc
 import DialPadIcon from "@vector-im/compound-design-tokens/assets/web/icons/dial-pad";
 
 import styles from "./RoomListSearchView.module.css";
-import { type ViewModel } from "../../viewmodel/ViewModel";
-import { useViewModel } from "../../useViewModel";
+import { type ViewModel, useViewModel } from "../../viewmodel";
 import { Flex } from "../../utils/Flex";
 import { useI18n } from "../../utils/i18nContext";
