+++ packages/shared-components/src/viewmodel/index.ts
@@ -12,3 +12,5 @@ export * from "./ViewModelSubscriptions";
 export type * from "./ViewModel";
 export * from "./MockViewModel";
 export * from "./useCreateAutoDisposedViewModel";
+export * from "./useViewModel";
+export * from "./useMockedViewModel";
